+++ deepinv/optim/prior.py
@@ -10,12 +10,12 @@ from deepinv.models.wavdict import WaveletDenoiser
 
 class Prior(nn.Module):
     r"""
-    Prior term :math:`g(x)`.
+    Prior term :math:`\reg{x}`.
 
-    This is the base class for the prior term :math:`g(x)`. Similarly to the :meth:`deepinv.optim.DataFidelity` class,
+    This is the base class for the prior term :math:`\reg{x}`. Similarly to the :meth:`deepinv.optim.DataFidelity` class,
     this class comes with methods for computing
-    :math:`\operatorname{prox}_{g}` and :math:`\nabla g`.
-    To implement a custom prior, for an explicit prior, overwrite :math:`g` (do not forget to specify
+    :math:`\operatorname{prox}_{g}` and :math:`\nabla \regname`.
+    To implement a custom prior, for an explicit prior, overwrite :math:`\regname` (do not forget to specify
     `self.explicit_prior = True`)
 
     This base class is also used to implement implicit priors. For instance, in PnP methods, the method computing the
@@ -58,7 +58,7 @@ class Prior(nn.Module):
 
     def grad(self, x, *args, **kwargs):
         r"""
-        Calculates the gradient of the prior term :math:`g` at :math:`x`.
+        Calculates the gradient of the prior term :math:`\regname` at :math:`x`.
         By default, the gradient is computed using automatic differentiation.
 
         :param torch.Tensor x: Variable :math:`x` at which the gradient is computed.
@@ -82,7 +82,7 @@ class Prior(nn.Module):
         **kwargs,
     ):
         r"""
-        Calculates the proximity operator of :math:`g` at :math:`x`. By default, the proximity operator is computed using internal gradient descent.
+        Calculates the proximity operator of :math:`\regname` at :math:`x`. By default, the proximity operator is computed using internal gradient descent.
 
         :param torch.Tensor x: Variable :math:`x` at which the proximity operator is computed.
         :param float gamma: stepsize of the proximity operator.
@@ -100,7 +100,7 @@ class Prior(nn.Module):
         r"""
         Calculates the proximity operator of the convex conjugate :math:`(\lambda g)^*` at :math:`x`, using the Moreau formula.
 
-        ::Warning:: Only valid for convex :math:`g`
+        ::Warning:: Only valid for convex :math:`\regname`
 
         :param torch.tensor x: Variable :math:`x` at which the proximity operator is computed.
         :param float gamma: stepsize of the proximity operator.
@@ -112,7 +112,7 @@ class Prior(nn.Module):
 
 class PnP(Prior):
     r"""
-    Plug-and-play prior :math:`\operatorname{prox}_{\gamma g}(x) = \operatorname{D}_{\sigma}(x)`.
+    Plug-and-play prior :math:`\operatorname{prox}_{\gamma \regname}(x) = \operatorname{D}_{\sigma}(x)`.
 
 
     :param callable denoiser: Denoiser :math:`\operatorname{D}_{\sigma}`.
@@ -125,7 +125,7 @@ class PnP(Prior):
 
     def prox(self, x, sigma_denoiser, *args, **kwargs):
         r"""
-        Uses denoising as the proximity operator of the PnP prior :math:`g` at :math:`x`.
+        Uses denoising as the proximity operator of the PnP prior :math:`\regname` at :math:`x`.
 
         :param torch.tensor x: Variable :math:`x` at which the proximity operator is computed.
         :param float sigma_denoiser: noise level parameter of the denoiser.
@@ -136,7 +136,7 @@ class PnP(Prior):
 
 class RED(Prior):
     r"""
-    Regularization-by-Denoising (RED) prior :math:`\nabla g(x) = \operatorname{Id} - \operatorname{D}_{\sigma}(x)`.
+    Regularization-by-Denoising (RED) prior :math:`\nabla \reg{x} = x - \operatorname{D}_{\sigma}(x)`.
 
 
     :param callable denoiser: Denoiser :math:`\operatorname{D}_{\sigma}`.
@@ -149,7 +149,7 @@ class RED(Prior):
 
     def grad(self, x, sigma_denoiser, *args, **kwargs):
         r"""
-        Calculates the gradient of the prior term :math:`g` at :math:`x`.
+        Calculates the gradient of the prior term :math:`\regname` at :math:`x`.
         By default, the gradient is computed using automatic differentiation.
 
         :param torch.Tensor x: Variable :math:`x` at which the gradient is computed.
@@ -160,7 +160,7 @@ class RED(Prior):
 
 class ScorePrior(Prior):
     r"""
-    Score via MMSE denoiser :math:`\nabla g(x)=\left(x-\operatorname{D}_{\sigma}(x)\right)/\sigma^2`.
+    Score via MMSE denoiser :math:`\nabla \reg{x}=\left(x-\operatorname{D}_{\sigma}(x)\right)/\sigma^2`.
 
     This approximates the score of a distribution using Tweedie's formula, i.e.,
 
@@ -208,7 +208,7 @@ class ScorePrior(Prior):
 
 class Tikhonov(Prior):
     r"""
-    Tikhonov regularizer :math:`g(x) = \frac{1}{2}\| x \|_2^2`.
+    Tikhonov regularizer :math:`\reg{x} = \frac{1}{2}\| x \|_2^2`.
     """
 
     def __init__(self, *args, **kwargs):
@@ -217,11 +217,11 @@ class Tikhonov(Prior):
 
     def g(self, x, ths=1.0):
         r"""
-        Computes the Tikhonov regularizer :math:`g(x) = \frac{\tau}{2}\| x \|_2^2`.
+        Computes the Tikhonov regularizer :math:`\reg{x} = \frac{\tau}{2}\| x \|_2^2`.
 
         :param torch.Tensor x: Variable :math:`x` at which the prior is computed.
         :param float ths: regularization parameter :math:`\tau`.
-        :return: (torch.Tensor) prior :math:`g(x)`.
+        :return: (torch.Tensor) prior :math:`\reg{x}`.
         """
         return (
             0.5
@@ -231,7 +231,7 @@ class Tikhonov(Prior):
 
     def grad(self, x):
         r"""
-        Calculates the gradient of the Tikhonov regularization term :math:`g` at :math:`x`.
+        Calculates the gradient of the Tikhonov regularization term :math:`\regname` at :math:`x`.
 
         :param torch.Tensor x: Variable :math:`x` at which the gradient is computed.
         :return: (torch.Tensor) gradient at :math:`x`.
@@ -252,7 +252,7 @@ class Tikhonov(Prior):
 
 class L1Prior(Prior):
     r"""
-    :math:`\ell_1` prior :math:`g(x) = \| x \|_1`.
+    :math:`\ell_1` prior :math:`\reg{x} = \| x \|_1`.
 
     """
 
@@ -262,17 +262,17 @@ class L1Prior(Prior):
 
     def g(self, x, ths=1.0):
         r"""
-        Computes the regularizer :math:`g(x) = \tau\| x \|_1`.
+        Computes the regularizer :math:`\reg{x} = \tau\| x \|_1`.
 
         :param torch.Tensor x: Variable :math:`x` at which the prior is computed.
         :param float ths: threshold parameter :math:`\tau`.
-        :return: (torch.Tensor) prior :math:`g(x)`.
+        :return: (torch.Tensor) prior :math:`\reg{x}`.
         """
         return ths * torch.norm(x.contiguous().view(x.shape[0], -1), p=1, dim=-1)
 
     def prox(self, x, ths=1.0, gamma=1.0):
         r"""
-        Calculates the proximity operator of the l1 regularization term :math:`g` at :math:`x`.
+        Calculates the proximity operator of the l1 regularization term :math:`\regname` at :math:`x`.
 
         More precisely, it computes
 
@@ -363,7 +363,7 @@ class WaveletPrior(Prior):
 
 class TVPrior(Prior):
     r"""
-    Total variation (TV) prior :math:`g(x) = \| D x \|_{1,2}`.
+    Total variation (TV) prior :math:`\reg{x} = \| D x \|_{1,2}`.
 
     :param float def_crit: default convergence criterion for the inner solver of the TV denoiser; default value: 1e-8.
     :param int n_it_max: maximal number of iterations for the inner solver of the TV denoiser; default value: 1000.