+++ deepinv/utils/plotting.py
@@ -74,7 +74,8 @@ def numpy2uint(img):
 
 def rescale_img(img, rescale_mode="min_max"):
     if rescale_mode == "min_max":
-        img = (img - img.min()) / (img.max() - img.min())
+        if img.max() != img.min():
+            img = (img - img.min()) / (img.max() - img.min())
     elif rescale_mode == "clip":
         img = img.clamp(min=0.0, max=1.0)
     else:
@@ -95,10 +96,15 @@ def plot(
     r"""
     Plots a list of images.
 
-    The images should be of shape [B,C,H,W], where B is the batch size, C is the number of channels,
+    The images should be of shape [B,C,H,W] or [C, H, W], where B is the batch size, C is the number of channels,
     H is the height and W is the width. The images are plotted in a grid, where the number of rows is B
     and the number of columns is the length of the list. If the B is bigger than max_imgs, only the first
     batches are plotted.
+
+    .. warning::
+
+        If the number of channels is 2, the magnitude of the complex images is plotted.
+        If the number of channels is bigger than 3, only the first 3 channels are plotted.
 
     Example usage:
 
@@ -127,6 +133,10 @@ def plot(
     if isinstance(img_list, torch.Tensor):
         img_list = [img_list]
 
+    for i, img in enumerate(img_list):
+        if len(img.shape) == 3:
+            img_list[i] = img.unsqueeze(0)
+
     if isinstance(titles, str):
         titles = [titles]
 
@@ -143,8 +153,13 @@ def plot(
                     .unsqueeze(0)
                     .type(torch.float32)
                 )
+            elif im.shape[1] > 3:
+                pimg = im[i, 0:3, :, :].type(torch.float32)
             else:
-                pimg = im[i, :, :, :].type(torch.float32)
+                if torch.is_complex(im):
+                    pimg = im[i, :, :, :].abs().type(torch.float32)
+                else:
+                    pimg = im[i, :, :, :].type(torch.float32)
             pimg = rescale_img(pimg, rescale_mode=rescale_mode)
             col_imgs.append(pimg.detach().permute(1, 2, 0).squeeze().cpu().numpy())
         imgs.append(col_imgs)