+++ CHANGELOG.rst
@@ -6,6 +6,22 @@ This change log is for the `main` branch. It contains changes for each release, 
 Current
 ----------------
 
+
+New Features
+^^^^^^^^^^^^
+
+
+Fixed
+^^^^^
+- Fixed g dependence on sigma in optim docs (:gh:`165` by `Julian Tachella`_) - 28/02/2024
+
+Changed
+^^^^^^^
+- Changed WaveletPrior to WaveletDenoiser (:gh:`165` by `Julian Tachella`_) - 28/02/2024
+
+v0.1.1
+----------------
+
 New Features
 ^^^^^^^^^^^^
 - Added r2r loss (:gh:`148` by `Brayan Monroy`_) - 30/01/2024
@@ -87,3 +103,4 @@ Authors
 .. _Riyad Chamekh: https://github.com/riyadchk
 .. _Jules Dumouchel: https://github.com/Ruli0
 .. _Brayan Monroy: https://github.com/bemc22
+.. _Nils Laurent: https://nils-laurent.github.io/